engine: Add probe subcommand
This provides an easy mechanism to check a simulation configuration
for available trigers, signals, etc.
