+++ engine/src/lua_api.hpp
@@ -36,7 +36,7 @@ namespace cloe {
  * Safely load and run a user Lua script.
  */
 [[nodiscard]] sol::protected_function_result lua_safe_script_file(
-    sol::state_view& lua, const std::filesystem::path& filepath);
+    sol::state_view lua, const std::filesystem::path& filepath);
 
 /**
  * Return the cloe-engine table as it is exported into Lua.
@@ -46,28 +46,28 @@ namespace cloe {
  *     engine/lua/cloe-engine/init.lua
  *
  */
-[[nodiscard]] inline auto luat_cloe_engine(sol::state_view& lua) {
-  return lua["package"]["loaded"]["cloe-engine"];
+[[nodiscard]] inline auto luat_cloe_engine(sol::state_view lua) {
+  return lua.globals()["package"]["loaded"]["cloe-engine"];
 }
 
-[[nodiscard]] inline auto luat_cloe_engine_fs(sol::state_view& lua) {
-  return lua["package"]["loaded"]["cloe-engine.fs"];
+[[nodiscard]] inline auto luat_cloe_engine_fs(sol::state_view lua) {
+  return lua.globals()["package"]["loaded"]["cloe-engine.fs"];
 }
 
-[[nodiscard]] inline auto luat_cloe_engine_types(sol::state_view& lua) {
-  return lua["package"]["loaded"]["cloe-engine.types"];
+[[nodiscard]] inline auto luat_cloe_engine_types(sol::state_view lua) {
+  return lua.globals()["package"]["loaded"]["cloe-engine.types"];
 }
 
-[[nodiscard]] inline auto luat_cloe_engine_initial_input(sol::state_view& lua) {
-  return lua["package"]["loaded"]["cloe-engine"]["initial_input"];
+[[nodiscard]] inline auto luat_cloe_engine_initial_input(sol::state_view lua) {
+  return lua.globals()["package"]["loaded"]["cloe-engine"]["initial_input"];
 }
 
-[[nodiscard]] inline auto luat_cloe_engine_state(sol::state_view& lua) {
-  return lua["package"]["loaded"]["cloe-engine"]["state"];
+[[nodiscard]] inline auto luat_cloe_engine_state(sol::state_view lua) {
+  return lua.globals()["package"]["loaded"]["cloe-engine"]["state"];
 }
 
-[[nodiscard]] inline auto luat_cloe_engine_plugins(sol::state_view& lua) {
-  return lua["package"]["loaded"]["cloe-engine"]["plugins"];
+[[nodiscard]] inline auto luat_cloe_engine_plugins(sol::state_view lua) {
+  return lua.globals()["package"]["loaded"]["cloe-engine"]["plugins"];
 }
 
 }  // namespace cloe
