+++ engine/src/simulation_context.hpp
@@ -22,210 +22,130 @@
 
 #pragma once
 
-#include <cstdint>                      // for uint64_t
-#include <functional>                   // for function<>
-#include <map>                          // for map<>
-#include <memory>                       // for unique_ptr<>, shared_ptr<>
-#include <optional>                     // for optional<>
-#include <string>                       // for string
-#include <vector>                       // for vector<>
+#include <functional>  // for function<>
+#include <map>         // for map<>
+#include <memory>      // for unique_ptr<>, shared_ptr<>
+#include <optional>    // for optional<>
+#include <string>      // for string
+#include <vector>      // for vector<>
 
-#include <sol/state_view.hpp>           // for state_view
+#include <sol/state_view.hpp>  // for state_view
 
-#include <cloe/cloe_fwd.hpp>            // for Simulator, Controller, Registrar, Vehicle, Duration
-#include <cloe/data_broker.hpp>         // for DataBroker
-#include <cloe/sync.hpp>                // for Sync
-#include <cloe/trigger/nil_event.hpp>   // for DEFINE_NIL_EVENT
-#include <cloe/utility/statistics.hpp>  // for Accumulator
-#include <cloe/utility/timer.hpp>       // for DurationTimer
+#include <cloe/cloe_fwd.hpp>       // for Simulator, Controller, Registrar, Vehicle, Duration
+#include <cloe/utility/timer.hpp>  // for DurationTimer
 
-#include "coordinator.hpp"              // for Coordinator
-#include "registrar.hpp"                // for Registrar
-#include "server.hpp"                   // for Server
-#include "simulation_progress.hpp"      // for SimulationProgress
-#include "stack.hpp"                    // for Stack
-#include "utility/command.hpp"          // for CommandExecuter
-#include "utility/time_event.hpp"       // for TimeCallback
+#include "simulation_events.hpp"      // for LoopCallback, ...
+#include "simulation_outcome.hpp"     // for SimulationOutcome
+#include "simulation_probe.hpp"       // for SimulationProbe
+#include "simulation_progress.hpp"    // for SimulationProgress
+#include "simulation_result.hpp"      // for SimulationResult
+#include "simulation_statistics.hpp"  // for SimulationStatistics
+#include "simulation_sync.hpp"        // for SimulationSync
+#include "stack.hpp"                  // for Stack
 
 namespace engine {
 
-/**
- * SimulationSync is the synchronization context of the simulation.
- */
-class SimulationSync : public cloe::Sync {
- public:  // Overrides
-  SimulationSync() = default;
-  explicit SimulationSync(const cloe::Duration& step_width) : step_width_(step_width) {}
-
-  uint64_t step() const override { return step_; }
-  cloe::Duration step_width() const override { return step_width_; }
-  cloe::Duration time() const override { return time_; }
-  cloe::Duration eta() const override { return eta_; }
-
-  /**
-   * Return the target simulation factor, with 1.0 being realtime.
-   *
-   * - If target realtime factor is <= 0.0, then it is interpreted to be unlimited.
-   * - Currently, the floating INFINITY value is not handled specially.
-   */
-  double realtime_factor() const override { return realtime_factor_; }
-
-  /**
-   * Return the maximum theorically achievable simulation realtime factor,
-   * with 1.0 being realtime.
-   */
-  double achievable_realtime_factor() const override {
-    return static_cast<double>(step_width().count()) / static_cast<double>(cycle_time_.count());
-  }
-
- public:  // Modification
-  /**
-   * Increase the step number for the simulation.
-   *
-   * - It increases the step by one.
-   * - It moves the simulation time forward by the step width.
-   * - It stores the real time difference from the last time IncrementStep was called.
-   */
-  void increment_step() {
-    step_ += 1;
-    time_ += step_width_;
-  }
-
-  /**
-   * Set the target realtime factor, with any value less or equal to zero
-   * unlimited.
-   */
-  void set_realtime_factor(double s) { realtime_factor_ = s; }
-
-  void set_eta(cloe::Duration d) { eta_ = d; }
-
-  void reset() {
-    time_ = cloe::Duration(0);
-    step_ = 0;
-  }
-
-  void set_cycle_time(cloe::Duration d) { cycle_time_ = d; }
-
- private:
-  // Simulation State
-  uint64_t step_{0};
-  cloe::Duration time_{0};
-  cloe::Duration eta_{0};
-  cloe::Duration cycle_time_;
-
-  // Simulation Configuration
-  double realtime_factor_{1.0};            // realtime
-  cloe::Duration step_width_{20'000'000};  // should be 20ms
-};
-
-struct SimulationStatistics {
-  cloe::utility::Accumulator engine_time_ms;
-  cloe::utility::Accumulator cycle_time_ms;
-  cloe::utility::Accumulator simulator_time_ms;
-  cloe::utility::Accumulator controller_time_ms;
-  cloe::utility::Accumulator padding_time_ms;
-  cloe::utility::Accumulator controller_retries;
-
-  void reset() {
-    engine_time_ms.reset();
-    cycle_time_ms.reset();
-    simulator_time_ms.reset();
-    controller_time_ms.reset();
-    padding_time_ms.reset();
-    controller_retries.reset();
-  }
-
-  friend void to_json(cloe::Json& j, const SimulationStatistics& s) {
-    j = cloe::Json{
-        {"engine_time_ms", s.engine_time_ms},         {"simulator_time_ms", s.simulator_time_ms},
-        {"controller_time_ms", s.controller_time_ms}, {"padding_time_ms", s.padding_time_ms},
-        {"cycle_time_ms", s.cycle_time_ms},           {"controller_retries", s.controller_retries},
-    };
-  }
-};
+// Forward-declarations:
+class CommandExecuter;
+class Registrar;
+class Coordinator;
+class Server;
+class SimulationResult;
+class SimulationProbe;
 
 /**
- * SimulationOutcome describes the possible outcomes a simulation can have.
- */
-enum class SimulationOutcome {
-  NoStart,  ///< Simulation unable to start.
-  Aborted,  ///< Simulation aborted due to technical problems or interrupt.
-  Stopped,  ///< Simulation concluded, but without valuation.
-  Failure,  ///< Simulation explicitly concluded with failure.
-  Success,  ///< Simulation explicitly concluded with success.
-};
-
-// If possible, the following exit codes should not be used as they are used
-// by the Bash shell, among others: 1-2, 126-165, and 255. That leaves us
-// primarily with the range 3-125, which should suffice for our purposes.
-// The following exit codes should not be considered stable.
-#define EXIT_OUTCOME_SUCCESS EXIT_SUCCESS  // normally 0
-#define EXIT_OUTCOME_UNKNOWN EXIT_FAILURE  // normally 1
-#define EXIT_OUTCOME_NOSTART 4             // 0b.....1..
-#define EXIT_OUTCOME_STOPPED 8             // 0b....1...
-#define EXIT_OUTCOME_FAILURE 9             // 0b....1..1
-#define EXIT_OUTCOME_ABORTED 16            // 0b...1....
-
-// clang-format off
-ENUM_SERIALIZATION(SimulationOutcome, ({
-    {SimulationOutcome::Aborted, "aborted"},
-    {SimulationOutcome::NoStart, "no-start"},
-    {SimulationOutcome::Failure, "failure"},
-    {SimulationOutcome::Success, "success"},
-    {SimulationOutcome::Stopped, "stopped"},
-}))
-// clang-format on
-
-namespace events {
-
-DEFINE_NIL_EVENT(Start, "start", "start of simulation")
-DEFINE_NIL_EVENT(Stop, "stop", "stop of simulation")
-DEFINE_NIL_EVENT(Success, "success", "simulation success")
-DEFINE_NIL_EVENT(Failure, "failure", "simulation failure")
-DEFINE_NIL_EVENT(Reset, "reset", "reset of simulation")
-DEFINE_NIL_EVENT(Pause, "pause", "pausation of simulation")
-DEFINE_NIL_EVENT(Resume, "resume", "resumption of simulation after pause")
-DEFINE_NIL_EVENT(Loop, "loop", "begin of inner simulation loop each cycle")
-
-}  // namespace events
-
-/**
- * SimulationContext represents the entire context of a running simulation.
+ * SimulationContext represents the entire context of a running simulation
+ * and is used by SimulationMachine class as the data context for the
+ * state machine.
  *
- * This clearly separates data from functionality. There is no constructor
- * where extra initialization is performed. Instead any initialization is
- * performed in the simulation states in the `simulation.cpp` file.
+ * The simulation states need to store any data they want to access in the
+ * context here. This does have the caveat that all the data here is
+ * accessible to all states.
+ *
+ * All input to and output from the simulation is via this struct.
  */
 struct SimulationContext {
-  SimulationContext(sol::state_view&& l) : lua(l) {}
+  SimulationContext(cloe::Stack conf, sol::state_view l);
 
+  // Configuration -----------------------------------------------------------
+  //
+  // These values are meant to be set before starting the simulation in order
+  // to affect how the simulation is run.
+  //
+  // The other values in this struct should not be directly modified unless
+  // you really know what you are doing.
+  //
+
+  cloe::Stack config;  ///< Input configuration.
+  std::string uuid{};  ///< UUID to use for simulation.
+
+  /// Report simulation progress to the console.
+  bool report_progress{false};
+
+  /// Setup simulation but only probe for information.
+  /// The simulation should only go through the CONNECT -> PROBE -> DISCONNECT
+  /// state. The same errors that can occur for a normal simulation can occur
+  /// here though, so make sure they are handled.
+  bool probe_simulation{false};
+
+  // Setup -------------------------------------------------------------------
+  //
+  // These are functional parts of the simulation framework that mostly come
+  // from the engine. They are all initialized in the constructor.
+  //
   sol::state_view lua;
-
-  // Setup
   std::unique_ptr<cloe::DataBroker> db;
   std::unique_ptr<Server> server;
   std::shared_ptr<Coordinator> coordinator;
   std::shared_ptr<Registrar> registrar;
+
+  /// Configurable system command executer for triggers.
   std::unique_ptr<CommandExecuter> commander;
 
-  // Configuration
-  cloe::Stack config;
-  std::string uuid{};
-  bool report_progress{false};
+  // State -------------------------------------------------------------------
+  //
+  // These are the types that represent the simulation state and have no
+  // functionality of their own, directly. They may change during the
+  // simulation.
+  //
 
-  // State
+  /// Track the simulation timing.
   SimulationSync sync;
+
+  /// Track the approximate progress of the simulation.
   SimulationProgress progress;
-  SimulationStatistics statistics;
+
+  /// Non-owning pointer used in order to keep track which model is being
+  /// initialized in the CONNECT state in order to allow it to be directly
+  /// aborted if it is hanging during initialization. If no model is being
+  /// actively initialized the valid value is nullptr.
   cloe::Model* now_initializing{nullptr};
+
   std::map<std::string, std::unique_ptr<cloe::Simulator>> simulators;
   std::map<std::string, std::shared_ptr<cloe::Vehicle>> vehicles;
   std::map<std::string, std::unique_ptr<cloe::Controller>> controllers;
-  std::optional<SimulationOutcome> outcome;
+
   timer::DurationTimer<cloe::Duration> cycle_duration;
+
+  /// Tell the simulation that we want to transition into the PAUSE state.
+  ///
+  /// We can't do this directly via an interrupt because we can only go
+  /// into the PAUSE state after STEP_END.
   bool pause_execution{false};
 
-  // Events
+  // Output ------------------------------------------------------------------
+  SimulationStatistics statistics;
+  std::optional<SimulationOutcome> outcome;
+  std::optional<SimulationResult> result;
+  std::optional<SimulationProbe> probe;
+
+  // Events ------------------------------------------------------------------
+  //
+  // The following callbacks store listeners on the given events.
+  // In the state where an event occurs, the callback is then triggered.
+  // There is generally only one place where each of these callbacks is
+  // triggered.
+  //
   std::shared_ptr<events::LoopCallback> callback_loop;
   std::shared_ptr<events::PauseCallback> callback_pause;
   std::shared_ptr<events::ResumeCallback> callback_resume;
@@ -237,8 +157,14 @@ struct SimulationContext {
   std::shared_ptr<events::TimeCallback> callback_time;
 
  public:
+  // Helper Methods ----------------------------------------------------------
+  //
+  // These methods encapsulate methods on the data in this struct that can be
+  // used by various states. They constitute implementation details and may
+  // be refactored out of this struct at some point.
+  //
   std::string version() const;
-  cloe::Logger logger() const { return cloe::logger::get("cloe"); }
+  cloe::Logger logger() const;
 
   std::shared_ptr<cloe::Registrar> simulation_registrar();
