+++ tests/test_lua14_speedometer_signals.lua
@@ -1,0 +1,7 @@
+local cloe = require("cloe")
+
+cloe.load_stackfile("config_minimator_multi_agent_infinite.json")
+
+local Sig = {
+    Speedometer = "third_speed/kmph"
+}
